lib: allow for redefinitions
This also adds support for forget. I originally tried handling forget
in tokenizer and eval_expr but found that it blew up the types.
Ultimately, forget and : need to be handled the same way and it
doesn't seem worth it right now to overhaul this thing into a proper
lexer and parser.