lib: define built-in words the same way as the others
This removes the special-casing of (most of) the built-in words,
unifying the lookup logic. This allows builtins to be overridden,
which should be fun. It also has the funny side effect of allowing
builtins to be forgotten, after which they cannot be redefined. The
construction of the initial dictionary was also cleaned up since it
started to get repetitive with all of the builtins.