lib: implement markers
This abuses the dictionary slightly to keep track of when markers were
defined. This has the benefit of allowing markers to be redefined,
just like words. This also nicely handles the case of nested markers -
when an "outer" one is executed, the "inner" ones are forgotten,
including the null definition.